# api/src/main/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Saamp Web API
  version: 1.0.0
...
paths:
...
  /api/users/{id}:
    get:
      tags: [Users]
      operationId: getUserById
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: './schemas/user.yaml#/User' } } } } # *
        '404': { description: Not found }

  # --- Endpoints Téléphonie ---
  /phone-logs/import:
    post:
      summary: Importer des fichiers CSV de logs téléphoniques
      operationId: importPhoneLogs
      tags: [PhoneLogs]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties: { files: { type: array, items: { type: string, format: binary } } }
              required: [ files ]
      responses:
        '200':
          description: Résultat de l'import
          content: { application/json: { schema: { $ref: './schemas/telephony.yaml#/PhoneLogImportResult' } } } # *
  /phone-logs/imports:
    get:
      tags: [PhoneLogs]
      summary: Récupérer l'historique des imports de fichiers
      operationId: getPhoneLogImports
      parameters:
        - { in: query, name: limit, schema: { type: integer, default: 10 } }
      responses:
        '200':
          description: Historique des imports
          content: { application/json: { schema: { type: array, items: { $ref: './schemas/telephony.yaml#/PhoneLogImportHistory' } } } } # *
...
  /phone-user/{phoneUserId}:
    get:
      tags: [ PhoneLogs ]
      summary: Details Phone User
      operationId: getPhoneUsersDetail
      parameters:
        - name: phoneUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Details Phone User
          content:
            application/json:
              schema: { $ref: './schemas/telephony.yaml#/PhoneUser' }
  /phone-users:
    get:
      tags: [ PhoneLogs ]
      summary: Liste paginée des phone users
      description: >
        Retourne la liste paginée des phone users avec possibilité de recherche
        sur nom, prénom, denomination, company et numéros de téléphone.
      operationId: getPhoneUsersPage
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Index de page (0-based)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 25
          description: Taille de page
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: >
            Texte de recherche (nom, prénom, denomination, company, téléphone)
      responses:
        "200":
          description: Page de phone users
          content:
            application/json:
              schema: { $ref: './schemas/telephony.yaml#/PhoneUserPage' }
  /phone-users/{mysaampIdClient}:
    get:
      tags: [ PhoneLogs ]
      summary: Liste paginée des phone users
      description: >
        Retourne la liste paginée des phone users avec possibilité de recherche
        sur nom, prénom, denomination, company et numéros de téléphone.
      operationId: getPhoneUsersPageWithClIdent
      parameters:
        - name: mysaampIdClient
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Index de page (0-based)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 25
          description: Taille de page
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: >
            Texte de recherche (nom, prénom, denomination, company, téléphone)
      responses:
        "200":
          description: Page de phone users
          content:
            application/json:
              schema: { $ref: './schemas/telephony.yaml#/PhoneUserPage' }
  /phone-users/{phoneUserId}/stats/ca-hm:
    get:
      tags: [ PhoneLogs ]
      operationId: getPhoneUserStats
      summary: CA_HM mensuel par phone user
      description:
        Retourne le cumul de CA_HM pour chaque phone user lié
        à un client MySAAMP sur un mois et une année donnés.
      parameters:
        - name: phoneUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: annee
          in: query
          required: true
          schema:
            type: integer
            example: 2025
        - name: mois
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 12
      responses:
        "200":
          description: Statistique mensuelle du phone user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneUserMonthlyCaHm"
  /phone-logs/dashboard:
    get:
      tags: [PhoneLogs]
      summary: Récupérer les KPIs et statistiques pour le dashboard
      operationId: getDashboardStats
      parameters:
        - { in: query, name: startDate, schema: { type: string, format: date-time } }
        - { in: query, name: endDate, schema: { type: string, format: date-time } }
        - { in: query, name: commercialId, schema: { type: string, format: uuid } }
        - { in: query, name: type, schema: { type: string, enum: [ ALL, APPEL, SMS ], default: ALL } }
      responses:
        '200':
          description: Statistiques agrégées
          content: { application/json: { schema: { $ref: './schemas/telephony.yaml#/DashboardStats' } } } # *
  /phone-logs/page:
    get:
      tags: [PhoneLogs]
      summary: Récupérer les logs téléphoniques paginés (recommandé)
      operationId: getPhoneLogsPage
      parameters:
        - { in: query, name: forfaitContactId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: startDate, schema: { type: string, format: date-time }, description: "Filtre période (optionnel)" }
        - { in: query, name: endDate, schema: { type: string, format: date-time }, description: "Filtre période (optionnel)" }
        - { in: query, name: page, schema: { type: integer, minimum: 0, default: 0 } }
        - { in: query, name: size, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: Page de logs téléphoniques
          content:
            application/json:
              schema: { $ref: './schemas/telephony.yaml#/PhoneLogPage' }
  /phone-logs/stats/correspondents:
    get:
      tags: [ PhoneLogs ]
      summary: Stats appels et SMS par correspondant (global), filtré par forfait et période
      operationId: getCorrespondentStats
      parameters:
        - { in: query, name: forfaitContactId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: startDate, schema: { type: string, format: date-time } }
        - { in: query, name: endDate, schema: { type: string, format: date-time } }
        - { in: query, name: limitCalls, schema: { type: integer, minimum: 1, maximum: 500, default: 50 } }
        - { in: query, name: limitSms, schema: { type: integer, minimum: 1, maximum: 500, default: 50 } }
      responses:
        '200':
          description: Stats par correspondant
          content:
            application/json:
              schema: { $ref: './schemas/telephony.yaml#/CorrespondentStatsResponse' }
  /phone/stats/sales-rep:
    get:
      tags: [PhoneLogs]
      summary: Indicateurs détaillés par commercial (KPIs + ratios + activité par heure)
      operationId: getSalesRepStats
      parameters:
        - { in: query, name: commercialId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: startDate, schema: { type: string, format: date-time } }
        - { in: query, name: endDate, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Statistiques du commercial sur la période
          content: { application/json: { schema: { $ref: './schemas/telephony.yaml#/SalesRepStats' } } }
...
  /pricing/rules:
    get:
      tags: [ PricingAdmin ]
      summary: (Étape 3) Récupérer les règles basées sur la Nature et le Type
      operationId: getRules
      parameters:
        - { in: query, name: natureId, required: true, schema: { type: string } }
        - { in: query, name: type, required: true, schema: { type: string, enum: [ FORFAIT, PRESTATION ] } }
      responses:
        '200':
          description: "Liste des règles correspondantes (ex: 'Or 60-100%')"
          content:
            application/json: { schema: { type: array, items: { $ref: './schemas/pricing.yaml#/TarifRule' } } }
  /pricing/grids/{ruleType}/{ruleCode}:
    get:
      tags: [ PricingAdmin ]
      summary: (Étape 4) Récupérer la grille de prix pour une règle spécifique
      operationId: getPriceGrid
      parameters:
        - { in: path, name: ruleType, required: true, schema: { type: string, enum: [ FORFAIT, PRESTATION ] } }
        - { in: path, name: ruleCode, required: true, schema: { type: string } }
      responses:
        '200':
          description: "La grille de prix éditable"
          content:
            application/json: { schema: { $ref: './schemas/pricing.yaml#/PriceGrid' } }
    put:
      tags: [ PricingAdmin ]
      summary: (Étape 4 - Sauvegarde) Mettre à jour une grille de prix
      operationId: updatePriceGrid
      parameters:
        - { in: path, name: ruleType, required: true, schema: { type: string, enum: [ FORFAIT, PRESTATION ] } }
        - { in: path, name: ruleCode, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json: { schema: { $ref: './schemas/pricing.yaml#/PriceGrid' } }
      responses:
        '200':
          description: "Grille mise à jour"
          content:
            application/json: { schema: { $ref: './schemas/pricing.yaml#/PriceGrid' } }
        '400':
          description: "Validation échouée (ex: chevauchement de poids)"
...
components:
  schemas:
...
    PhoneUserMonthlyCaHm: { $ref: './schemas/telephony.yaml#/PhoneUserMonthlyCaHm' }
...